Move the testing bits of Module to a testing mod
Moves `Module::get_tests()` and `Module::run_tests()` into a `testing`
submodule.
