+++ src/codegen/mod.rs
@@ -25,7 +25,7 @@ use crate::{
         RuntimeFunctionRef,
     },
     typechecker::{info::TypeInfo, scope::ScopeRef, types},
-    IrValue, Runtime, Verdict,
+    IrValue, Runtime,
 };
 use check::{
     check_roto_type_reflect, FunctionRetrievalError, RotoFunc, TypeMismatch,
@@ -50,10 +50,9 @@ use cranelift::{
 };
 use cranelift_codegen::ir::SigRef;
 use log::info;
-pub use testcase::TestCase;
 
 pub mod check;
-mod testcase;
+pub mod testing;
 #[cfg(test)]
 mod tests;
 
@@ -1147,66 +1146,6 @@ impl<'c> FuncGen<'c> {
 }
 
 impl Module {
-    pub fn get_tests<Ctx: 'static>(
-        &mut self,
-    ) -> impl Iterator<Item = TestCase<Ctx>> + use<'_, Ctx> {
-        let tests: Vec<_> = self
-            .functions
-            .keys()
-            .filter(|x| {
-                x.rsplit_once(".")
-                    .map_or(x.as_ref(), |x| x.1)
-                    .starts_with("test#")
-            })
-            .map(Clone::clone)
-            .collect();
-
-        tests.into_iter().map(|name| {
-            TestCase::new(
-                name.replace("test#", ""),
-                self.get_function::<Ctx, fn() -> Verdict<(), ()>>(
-                    name.strip_prefix("pkg.").unwrap(),
-                )
-                .unwrap(),
-            )
-        })
-    }
-
-    pub fn run_tests<Ctx: 'static>(
-        &mut self,
-        mut ctx: Ctx,
-    ) -> Result<(), ()> {
-        let tests: Vec<_> = self.get_tests::<Ctx>().collect();
-
-        let total = tests.len();
-        let total_width = total.to_string().len();
-        let mut successes = 0;
-        let mut failures = 0;
-
-        for (n, test) in tests.into_iter().enumerate() {
-            let n = n + 1;
-            let test_display = test.name();
-            print!("Test {n:>total_width$} / {total}: {test_display}... ");
-
-            if test.run(&mut ctx) == Ok(()) {
-                successes += 1;
-                println!("\x1B[92mok\x1B[m");
-            } else {
-                failures += 1;
-                println!("\x1B[91mfail\x1B[m");
-            }
-        }
-        println!(
-            "Ran {total} tests, {successes} succeeded, {failures} failed"
-        );
-
-        if failures == 0 {
-            Result::Ok(())
-        } else {
-            Result::Err(())
-        }
-    }
-
     pub fn get_function<Ctx: 'static, F: RotoFunc>(
         &mut self,
         name: &str,
